updatecheck: move up for others to imports
