+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -231,7 +231,6 @@ go_library(
         "//cmd/frontend/hooks",
         "//cmd/frontend/hubspot",
         "//cmd/frontend/hubspot/hubspotutil",
-        "//cmd/frontend/internal/app/updatecheck",
         "//cmd/frontend/internal/processrestart",
         "//cmd/frontend/internal/search/logs",
         "//enterprise/cmd/worker/shared/sourcegraphoperator",
@@ -320,6 +319,7 @@ go_library(
         "//internal/txemail",
         "//internal/txemail/txtypes",
         "//internal/types",
+        "//internal/updatecheck",
         "//internal/usagestats",
         "//internal/users",
         "//internal/version",
