+++ cmd/frontend/internal/cli/BUILD.bazel
@@ -27,7 +27,6 @@ go_library(
         "//cmd/frontend/internal/app",
         "//cmd/frontend/internal/app/assetsutil",
         "//cmd/frontend/internal/app/ui",
-        "//cmd/frontend/internal/app/updatecheck",
         "//cmd/frontend/internal/bg",
         "//cmd/frontend/internal/cli/middleware",
         "//cmd/frontend/internal/httpapi",
@@ -77,6 +76,7 @@ go_library(
         "//internal/sysreq",
         "//internal/trace",
         "//internal/types",
+        "//internal/updatecheck",
         "//internal/users",
         "//internal/version",
         "//internal/version/upgradestore",
