+++ cmd/frontend/internal/httpapi/BUILD.bazel
@@ -27,7 +27,6 @@ go_library(
         "//cmd/frontend/envvar",
         "//cmd/frontend/globals",
         "//cmd/frontend/graphqlbackend",
-        "//cmd/frontend/internal/app/updatecheck",
         "//cmd/frontend/internal/handlerutil",
         "//cmd/frontend/internal/httpapi/releasecache",
         "//cmd/frontend/internal/httpapi/router",
@@ -64,6 +63,7 @@ go_library(
         "//internal/txemail",
         "//internal/txemail/txtypes",
         "//internal/types",
+        "//internal/updatecheck",
         "//internal/usagestats",
         "//lib/errors",
         "//schema",
